projects_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
requests 
PyJWT>=2.0.0
python-dotenv>=0.19.0